+++ src/ArtifactsMMO.NET/Endpoints/Characters/Characters.cs
@@ -1,4 +1,5 @@
 ﻿using ArtifactsMMO.NET.Enums.ErrorCodes.Character;
+using ArtifactsMMO.NET.Internal;
 using ArtifactsMMO.NET.Objects;
 using ArtifactsMMO.NET.Objects.Characters;
 using ArtifactsMMO.NET.Queries;
@@ -15,6 +16,12 @@ namespace ArtifactsMMO.NET.Endpoints.Characters
     {
         private readonly IValidator<string> _nameValidator = new CharacterNameValidator();
         public Characters(HttpClient httpClient, string apiKey) : base(httpClient, apiKey)
+        {
+        }
+
+        internal Characters(HttpClient httpClient, string apiKey,
+            IJsonSerializerOptionsFactory jsonSerializerOptionsFactory)
+            : base(httpClient, apiKey, jsonSerializerOptionsFactory)
         {
         }