+++ src/ArtifactsMMO.NET/Internal/IJsonSerializerOptionsFactory.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json;
+
+namespace ArtifactsMMO.NET.Internal
+{
+    internal interface IJsonSerializerOptionsFactory
+    {
+        JsonSerializerOptions Get(JsonSerializerOptionsMode mode);
+    }
+}